Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/css/css-text/white-space/seg-break-transformation-004-expected.txt


FAIL linebreak only ₩２４ assert_equals: expected 73 but got 80
FAIL spaces linebreak ₩２４ assert_equals: expected 73 but got 80
FAIL linebreak spaces ₩２４ assert_equals: expected 73 but got 80
FAIL spaces linebreak spaces ₩２４ assert_equals: expected 73 but got 80
FAIL multiple linebreaks ₩２４ assert_equals: expected 73 but got 80
FAIL multiple linebreaks + spaces ₩２４ assert_equals: expected 73 but got 80
FAIL linebreak only ２４₩ assert_equals: expected 73 but got 80
FAIL spaces linebreak ２４₩ assert_equals: expected 73 but got 80
FAIL linebreak spaces ２４₩ assert_equals: expected 73 but got 80
FAIL spaces linebreak spaces ２４₩ assert_equals: expected 73 but got 80
FAIL multiple linebreaks ２４₩ assert_equals: expected 73 but got 80
FAIL multiple linebreaks + spaces ２４₩ assert_equals: expected 73 but got 80
₩ ２４
₩ ２４
₩ ２４
₩ ２４
₩ ２４
₩ ２４
₩２４
２４ ₩
２４ ₩
２４ ₩
２４ ₩
２４ ₩
２４ ₩
２４₩
